feat(paginationAndFilterMonthAndWeek): adding pagination and filter and event-type body

+++ plugins/xpeapp-backend/src/agenda/events_types/delete_events_types.php
@@ -12,23 +12,34 @@ function apiDeleteEventsTypes(WP_REST_Request $request)
     // Get the id from the request body
     $id = $request->get_param('id');
 
+    $response = null;
+
     if (empty($id)) {
-        return new WP_Error('missing_params', __('Missing id', 'Agenda'), array('status' => 400));
+        $response = createErrorResponse('missing_params', 'Missing id', 400);
+    } else {
+        // Check if an event has this type
+        $table_events = $wpdb->prefix . 'agenda_events';
+        $exists = $wpdb->get_var($wpdb->prepare("SELECT COUNT(*) FROM $table_events WHERE type_id = %d", intval($id)));
+        if ($exists > 0) {
+            $response = createErrorResponse('event_exists', 'Cannot delete because event type has assigned events', 409);
+        } else {
+            // Delete the event type from the database
+            $result = $wpdb->delete(
+                $table_events_type,
+                array(
+                    'id' => intval($id)
+                )
+            );
+
+            // Check if the delete was successful
+            if ($result === false) {
+                $response = createErrorResponse('db_delete_error', 'Could not delete event type', 500);
+            } else {
+                // Return a 204 response
+                $response = createSuccessResponse('Event type deleted', 204);
+            }
+        }
     }
 
-    // Delete the event type from the database
-    $result = $wpdb->delete(
-        $table_events_type,
-        array(
-            'id' => intval($id)
-        )
-    );
-
-    // Check if the delete was successful
-    if ($result === false) {
-        return new WP_Error('db_delete_error', __('Could not delete event type', 'Agenda'), array('status' => 500));
-    }
-
-    // Return a 204 response
-    return new WP_REST_Response(null, 204);
+    return $response;
 }
